Huge refactor of the database querying code.
Falling back to SQL wasn't going to cut it where we wanted to use
end-chain AR methods like .pluck on our queries. The solution to
keep it AR all the way along was to create an anonymous class for
each thread so that it can modify the connection on the class and
not affect the other threads. Cheers @workmad3 for some good tips
in #rubyonrails :+1:.

Also added the block syntax which yields the scope results so you
can do additional stuff to each result set before they're reduced:

    dbs = ["customer1", "customer2"]
    scope = Brand.where('name like "%Ad"')
    results = Querrel.new(dbs).query(scope) do |s|
      s.pluck(:name)
    end
    puts results.inspect
    # => ["Addidas", "Addidas Originals", "Addidas", etc]

TODO: Add thread pooling so we don't have n threads all the time
and fix warning ("loading in progress, circular require considered
harmful")